Validation fix for multiple char
The validation for num_page_entry box for only integer entries used to
validate for single characters given as input so when ever a string
starting with an integer is copied with Ctrl+V it is validated.For
eg "12ABCD".

Now checking for each char of the string before validating.